Fixing stuffs

+++ MagicMirror.ConsoleApp/Models/MainViewModel.cs
@@ -20,7 +20,7 @@
 
         public string TravelTime { get; set; }
 
-        public int Distance { get; set; }
+        public double Distance { get; set; }
 
         public string DistanceUom { get; set; }
     }
